New features added

+++ src/compensation/compensation.service.ts
@@ -4,8 +4,8 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
 import { UpdateCompensationDto } from './dto/update-compensation.dto';
-import { Compensation } from './compensatio.entity';
-import { CreateCompensationDto } from './dto/create-compensatio.dto';
+import { Compensation } from './compensation.entity';
+import { CreateCompensationDto } from './dto/create-compensation.dto';
 
 @Injectable()
 export class CompensationService {
@@ -14,8 +14,19 @@ export class CompensationService {
     private compensationRepository: Repository<Compensation>,
   ) {}
 
-  async create(createCompensationDto: CreateCompensationDto): Promise<Compensation> {
-    const compensation = this.compensationRepository.create(createCompensationDto);
+  async create(createDto: CreateCompensationDto): Promise<Compensation> {
+
+    // Costo del combustible
+    const fuelCost = (createDto.distance / createDto.kmPerLitre) * createDto.fuelPrice;
+    const totalAmount = (fuelCost + 1200) / createDto.availableSeats;
+    
+    const compensation = this.compensationRepository.create({
+      idTrip: createDto.idTrip, // Asociando viaje a la compensacion
+      amount: totalAmount,
+      ratePerKm: fuelCost / createDto.distance,
+      createdAt: new Date(),
+      active: true,
+    });
     return this.compensationRepository.save(compensation);
   }
 
@@ -24,7 +35,10 @@ export class CompensationService {
   }
 
   async findOne(id: number): Promise<Compensation> {
-    const compensation = await this.compensationRepository.findOne({ where: { id } });
+    const compensation = await this.compensationRepository.findOne({
+      where: { idCompensation: id },
+    });
+  
     if (!compensation) {
       throw new NotFoundException(`Compensation with ID ${id} not found`);
     }
@@ -32,16 +46,30 @@ export class CompensationService {
   }
 
   async update(id: number, updateCompensationDto: UpdateCompensationDto): Promise<Compensation> {
-    await this.compensationRepository.update(id, updateCompensationDto);
-    const updatedCompensation = await this.compensationRepository.findOne({ where: { id } });
-    if (!updatedCompensation) {
-      throw new NotFoundException(`Compensation with ID ${id} not found`);
-    }
-    return updatedCompensation;
+    const compensation = await this.findOne(id);
+
+    Object.assign(compensation, updateCompensationDto);
+
+    return this.compensationRepository.save(compensation);
   }
 
   async remove(id: number): Promise<void> {
     const compensation = await this.findOne(id);
+    if (!compensation) {
+      throw new NotFoundException({
+        statusCode: 404,
+        errorCode: 'COMPENSATION_NOT_FOUND',
+        message: 'La compensación no existe o los datos no coinciden',
+      });
+    }
     await this.compensationRepository.remove(compensation);
+  }
+
+  async calculateCompensation(createCompensationDto: CreateCompensationDto) {
+    const { distance, kmPerLitre, availableSeats } = createCompensationDto;
+    const priceBusTicket = 1200;
+    const fuelCost = (distance / kmPerLitre) * 1000;
+    const totalAmount = (fuelCost + priceBusTicket) / availableSeats;
+    return { totalAmount };
   }
 }
